refactor message testing
